issue 8 cardのデータがない場合は 「データがありません」と表示する

+++ intern-book-app/src/app/components/bookinformation/bookinformation.component.ts
@@ -32,7 +32,9 @@ export class BookinformationComponent {
   constructor(private bookService:BookService){}
   
   deleteBook(book:any){
+    if(confirm("「本名」を削除します。よろしいですか？　")){
     this.bookService.deleteBook(book);
+    }
     
   }
 }
